requirements.txt
cryptography==42.0.8
matplotlib==3.5.3
numpy==1.21.6
nvidia_ml_py==12.535.161
nvidia_ml_py3==7.352.0
onnxruntime==1.14.1
onnxruntime_gpu==1.14.1
opencv_python==4.9.0.80
pandas==1.3.5
Pillow==9.5.0
Pillow==10.3.0
scikit_learn==1.0.2
scipy==1.7.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.2
